:hammer: [chore] SettingView 추가 #12
